Refactoring and creating the CMS data structure

+++ synapse/src/targets/tofino/data_structures/meter.cpp
@@ -4,7 +4,7 @@
 
 namespace tofino {
 
-Meter::Meter(DS_ID _id, u64 _capacity, Bps_t _rate, bytes_t _burst,
+Meter::Meter(DS_ID _id, u32 _capacity, Bps_t _rate, bytes_t _burst,
              const std::vector<bits_t> &_keys)
     : DS(DSType::METER, true, _id), capacity(_capacity), rate(_rate),
       burst(_burst), keys(_keys) {
