+++ synapse/src/targets/tofino/data_structures/meter.h
@@ -12,12 +12,12 @@
 namespace tofino {
 
 struct Meter : public DS {
-  u64 capacity;
+  u32 capacity;
   Bps_t rate;
   bytes_t burst;
   std::vector<bits_t> keys;
 
-  Meter(DS_ID id, u64 capacity, Bps_t rate, bytes_t burst,
+  Meter(DS_ID id, u32 capacity, Bps_t rate, bytes_t burst,
         const std::vector<bits_t> &keys);
 
   Meter(const Meter &other);
